Import parseEmbedMode.
